+++ Fantasy.Net/Fantasy.Net/Runtime/Core/DataStructure/NativeCollections/NativeCollections/NativeConcurrentStack.cs
@@ -12,12 +12,7 @@ using System.Threading;
 #pragma warning disable CA2208
 #pragma warning disable CS8632
 
-// ReSharper disable ConvertToAutoProperty
-// ReSharper disable ConvertToAutoPropertyWhenPossible
-// ReSharper disable ConvertToAutoPropertyWithPrivateSetter
-// ReSharper disable ConvertIfStatementToSwitchStatement
-// ReSharper disable PossibleNullReferenceException
-// ReSharper disable MemberHidesStaticFromOuterClass
+// ReSharper disable ALL
 
 namespace NativeCollections
 {
@@ -65,7 +60,7 @@ namespace NativeCollections
         public NativeConcurrentStack(int size, int maxFreeSlabs)
         {
             var nodePool = new NativeMemoryPool(size, sizeof(Node), maxFreeSlabs);
-            _handle = (NativeConcurrentStackHandle*)NativeMemoryAllocator.Alloc(sizeof(NativeConcurrentStackHandle));
+            _handle = (NativeConcurrentStackHandle*)NativeMemoryAllocator.Alloc((uint)sizeof(NativeConcurrentStackHandle));
             _handle->Head = IntPtr.Zero;
             _handle->NodePool = nodePool;
             _handle->NodePoolLock = new NativeConcurrentSpinLock(-1);
