+++ Fantasy.Net/Fantasy.Net/Runtime/Core/DataStructure/NativeCollections/NativeCollections/NativeQueue.cs
@@ -7,12 +7,7 @@ using System;
 #pragma warning disable CA2208
 #pragma warning disable CS8632
 
-// ReSharper disable ConvertToAutoProperty
-// ReSharper disable ConvertToAutoPropertyWhenPossible
-// ReSharper disable ConvertToAutoPropertyWithPrivateSetter
-// ReSharper disable ConvertIfStatementToSwitchStatement
-// ReSharper disable PossibleNullReferenceException
-// ReSharper disable MemberHidesStaticFromOuterClass
+// ReSharper disable ALL
 
 namespace NativeCollections
 {
@@ -76,8 +71,8 @@ namespace NativeCollections
                 throw new ArgumentOutOfRangeException(nameof(capacity), capacity, "MustBeNonNegative");
             if (capacity < 4)
                 capacity = 4;
-            _handle = (NativeQueueHandle*)NativeMemoryAllocator.Alloc(sizeof(NativeQueueHandle));
-            _handle->Array = (T*)NativeMemoryAllocator.Alloc(capacity * sizeof(T));
+            _handle = (NativeQueueHandle*)NativeMemoryAllocator.Alloc((uint)sizeof(NativeQueueHandle));
+            _handle->Array = (T*)NativeMemoryAllocator.Alloc((uint)(capacity * sizeof(T)));
             _handle->Length = capacity;
             _handle->Head = 0;
             _handle->Tail = 0;
@@ -318,7 +313,7 @@ namespace NativeCollections
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         private void SetCapacity(int capacity)
         {
-            var newArray = (T*)NativeMemoryAllocator.Alloc(capacity * sizeof(T));
+            var newArray = (T*)NativeMemoryAllocator.Alloc((uint)(capacity * sizeof(T)));
             if (_handle->Size > 0)
             {
                 if (_handle->Head < _handle->Tail)
