+++ Fantasy.Net/Fantasy.Net/Runtime/Core/DataStructure/NativeCollections/NativeCollections/NativeSortedDictionary.cs
@@ -10,12 +10,7 @@ using System.Runtime.InteropServices;
 #pragma warning disable CA2208
 #pragma warning disable CS8632
 
-// ReSharper disable ConvertToAutoProperty
-// ReSharper disable ConvertToAutoPropertyWhenPossible
-// ReSharper disable ConvertToAutoPropertyWithPrivateSetter
-// ReSharper disable ConvertIfStatementToSwitchStatement
-// ReSharper disable PossibleNullReferenceException
-// ReSharper disable MemberHidesStaticFromOuterClass
+// ReSharper disable ALL
 
 namespace NativeCollections
 {
@@ -87,7 +82,7 @@ namespace NativeCollections
         public NativeSortedDictionary(int size, int maxFreeSlabs)
         {
             var nodePool = new NativeMemoryPool(size, sizeof(Node), maxFreeSlabs);
-            _handle = (NativeSortedDictionaryHandle*)NativeMemoryAllocator.Alloc(sizeof(NativeSortedDictionaryHandle));
+            _handle = (NativeSortedDictionaryHandle*)NativeMemoryAllocator.Alloc((uint)sizeof(NativeSortedDictionaryHandle));
             _handle->Root = null;
             _handle->Count = 0;
             _handle->Version = 0;
