+++ Fantasy.Net/Fantasy.Net/Runtime/Core/DataStructure/NativeCollections/NativeCollections/NativeStack.cs
@@ -7,12 +7,7 @@ using System;
 #pragma warning disable CA2208
 #pragma warning disable CS8632
 
-// ReSharper disable ConvertToAutoProperty
-// ReSharper disable ConvertToAutoPropertyWhenPossible
-// ReSharper disable ConvertToAutoPropertyWithPrivateSetter
-// ReSharper disable ConvertIfStatementToSwitchStatement
-// ReSharper disable PossibleNullReferenceException
-// ReSharper disable MemberHidesStaticFromOuterClass
+// ReSharper disable ALL
 
 namespace NativeCollections
 {
@@ -66,8 +61,8 @@ namespace NativeCollections
                 throw new ArgumentOutOfRangeException(nameof(capacity), capacity, "MustBeNonNegative");
             if (capacity < 4)
                 capacity = 4;
-            _handle = (NativeStackHandle*)NativeMemoryAllocator.Alloc(sizeof(NativeStackHandle));
-            _handle->Array = (T*)NativeMemoryAllocator.Alloc(capacity * sizeof(T));
+            _handle = (NativeStackHandle*)NativeMemoryAllocator.Alloc((uint)sizeof(NativeStackHandle));
+            _handle->Array = (T*)NativeMemoryAllocator.Alloc((uint)(capacity * sizeof(T)));
             _handle->Length = capacity;
             _handle->Size = 0;
             _handle->Version = 0;
@@ -317,7 +312,7 @@ namespace NativeCollections
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         private void SetCapacity(int capacity)
         {
-            var newArray = (T*)NativeMemoryAllocator.Alloc(capacity * sizeof(T));
+            var newArray = (T*)NativeMemoryAllocator.Alloc((uint)(capacity * sizeof(T)));
             if (_handle->Size > 0)
                 Unsafe.CopyBlockUnaligned(newArray, _handle->Array, (uint)(_handle->Length * sizeof(T)));
             NativeMemoryAllocator.Free(_handle->Array);
