+++ Fantasy.Unity/Fantasy.Unity/Runtime/Core/DataStructure/NativeCollections/NativeCollections/NativeHashSet.cs
@@ -7,12 +7,7 @@ using System;
 #pragma warning disable CA2208
 #pragma warning disable CS8632
 
-// ReSharper disable ConvertToAutoProperty
-// ReSharper disable ConvertToAutoPropertyWhenPossible
-// ReSharper disable ConvertToAutoPropertyWithPrivateSetter
-// ReSharper disable ConvertIfStatementToSwitchStatement
-// ReSharper disable PossibleNullReferenceException
-// ReSharper disable MemberHidesStaticFromOuterClass
+// ReSharper disable ALL
 
 namespace NativeCollections
 {
@@ -91,7 +86,7 @@ namespace NativeCollections
                 throw new ArgumentOutOfRangeException(nameof(capacity), capacity, "MustBeNonNegative");
             if (capacity < 4)
                 capacity = 4;
-            _handle = (NativeHashSetHandle*)NativeMemoryAllocator.Alloc(sizeof(NativeHashSetHandle));
+            _handle = (NativeHashSetHandle*)NativeMemoryAllocator.Alloc((uint)sizeof(NativeHashSetHandle));
             _handle->Count = 0;
             _handle->FreeCount = 0;
             _handle->Version = 0;
@@ -370,8 +365,8 @@ namespace NativeCollections
         {
             var size = HashHelpers.GetPrime(capacity);
             _handle->FreeList = -1;
-            _handle->Buckets = (int*)NativeMemoryAllocator.AllocZeroed(size * sizeof(int));
-            _handle->Entries = (Entry*)NativeMemoryAllocator.AllocZeroed(size * sizeof(Entry));
+            _handle->Buckets = (int*)NativeMemoryAllocator.AllocZeroed((uint)(size * sizeof(int)));
+            _handle->Entries = (Entry*)NativeMemoryAllocator.AllocZeroed((uint)(size * sizeof(Entry)));
             _handle->BucketsLength = size;
             _handle->EntriesLength = size;
             _handle->FastModMultiplier = IntPtr.Size == 8 ? HashHelpers.GetFastModMultiplier((uint)size) : 0;
@@ -391,10 +386,10 @@ namespace NativeCollections
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         private void Resize(int newSize)
         {
-            var entries = (Entry*)NativeMemoryAllocator.AllocZeroed(newSize * sizeof(Entry));
+            var entries = (Entry*)NativeMemoryAllocator.AllocZeroed((uint)(newSize * sizeof(Entry)));
             var count = _handle->Count;
             Unsafe.CopyBlockUnaligned(entries, _handle->Entries, (uint)(_handle->EntriesLength * sizeof(Entry)));
-            var buckets = (int*)NativeMemoryAllocator.AllocZeroed(newSize * sizeof(int));
+            var buckets = (int*)NativeMemoryAllocator.AllocZeroed((uint)(newSize * sizeof(int)));
             NativeMemoryAllocator.Free(_handle->Buckets);
             _handle->Buckets = buckets;
             _handle->BucketsLength = newSize;
