+++ Fantasy.Unity/Fantasy.Unity/Runtime/Core/DataStructure/NativeCollections/NativeCollections/NativeMemoryBucket.cs
@@ -7,12 +7,7 @@ using System;
 #pragma warning disable CA2208
 #pragma warning disable CS8632
 
-// ReSharper disable ConvertToAutoProperty
-// ReSharper disable ConvertToAutoPropertyWhenPossible
-// ReSharper disable ConvertToAutoPropertyWithPrivateSetter
-// ReSharper disable ConvertIfStatementToSwitchStatement
-// ReSharper disable PossibleNullReferenceException
-// ReSharper disable MemberHidesStaticFromOuterClass
+// ReSharper disable ALL
 
 namespace NativeCollections
 {
@@ -71,10 +66,10 @@ namespace NativeCollections
                 throw new ArgumentOutOfRangeException(nameof(size), size, "MustBePositive");
             if (length < 0)
                 throw new ArgumentOutOfRangeException(nameof(length), length, "MustBeNonNegative");
-            _handle = (NativeMemoryBucketHandle*)NativeMemoryAllocator.Alloc(sizeof(NativeMemoryBucketHandle));
+            _handle = (NativeMemoryBucketHandle*)NativeMemoryAllocator.Alloc((uint)sizeof(NativeMemoryBucketHandle));
             _handle->Size = size;
             _handle->Length = length;
-            _handle->Array = (void**)NativeMemoryAllocator.AllocZeroed(size * sizeof(void*));
+            _handle->Array = (void**)NativeMemoryAllocator.AllocZeroed((uint)(size * sizeof(void*)));
             _handle->Index = 0;
             _handle->MemoryPool = new NativeMemoryPool(size, length, 0);
         }
